+++ common/src/main/java/io/github/mortuusars/exposure/util/ColorUtils.java
@@ -1,5 +1,7 @@
 package io.github.mortuusars.exposure.util;
 
+import java.awt.*;
+
 public class ColorUtils {
     public static int BGRtoRGB(int bgr) {
         int a = (bgr >> 24) & 0xFF;
@@ -8,5 +10,9 @@ public class ColorUtils {
         int r = bgr & 0xFF;
 
         return a << 24 | r << 16 | g << 8 | b;
+    }
+
+    public static int getRGBFromHex(String hexColor) {
+        return new Color((int) Long.parseLong(hexColor.replace("#", ""), 16), true).getRGB();
     }
 }