+++ fabric/src/main/java/io/github/mortuusars/exposure/fabric/resources/FabricFiltersLoader.java
@@ -1,12 +1,11 @@
 package io.github.mortuusars.exposure.fabric.resources;
 
 import io.github.mortuusars.exposure.Exposure;
-import io.github.mortuusars.exposure.data.FiltersResourceLoader;
-import io.github.mortuusars.exposure.data.LensesDataLoader;
+import io.github.mortuusars.exposure.data.filter.Filters;
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
 import net.minecraft.resources.ResourceLocation;
 
-public class FabricFiltersResourceLoader extends FiltersResourceLoader implements IdentifiableResourceReloadListener {
+public class FabricFiltersLoader extends Filters.Loader implements IdentifiableResourceReloadListener {
     public static final ResourceLocation ID = Exposure.resource("filters_loader");
     @Override
     public ResourceLocation getFabricId() {