+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Photographs can no longer be placed on the wall. Existing hanging photographs will not be removed.
 - Added Film Roll renaming. Use the item to open renaming UI. Renaming this way does not cost any experience.
 
+- Added ability to render custom Filter texture in Camera Attachments UI `//WIKI`
+
 - Lightroom comparator output now based on the number of Photographs in the output slot rather than the selected frame.
 - Chromatic Process can now be enabled for all black and white images by placing an Amethyst Cluster on top of a Lightroom.
 - Color Film recipe now requires Gold ingots and nuggets instead of Iron.